Add unreachable code rule
This adds a new rule that detect unreachable code, currently limited to
function bodies.

How it Works
============

The rule works as follows.

First we create "basic blocks" from the statements. These basic block
are zero or more lines of code (statements) for which the code flow is
easy to follow. Specifically all statements in a single block follow
each other, no diversion of the control flow. At the end of the block
the control can do one of three things: 1) continue to another code
block, 2) based on a condition jump to one of two code blocks, or 3)
terminate (return or end of the function).

Second, based on these basic blocks, and the simplified control flow
they represent, we can determine what blocks are and aren't reached. We
do this by starting with the first block of the function and following
the jumps it makes to the next code block, marking them all as reached.

Third, we create a diagnostic for each code block that is not reached by
step 2. Currently these diagnostics are quite limited, see below.

Future Work and Limitations
===========================

This commit is only the beginning of this rule, there is much work still
left to do.

The diagnostics created currently is quite limited. It only mentions the
function name and points to the fist statement in the basic block. In
the future this should be expanded to point to all statements in the
basic block. Furthermore it would be helpful to the users to explain
*why* a code block is not reached as currently, except for the most
basic example, this might not be clear.

We're currently quite limited on how we determine if a branch is always
taken or not, specifically we only detect the constants true and false
and only in the most basic condition (mainly the `if` and `while`
statements). This can be expanded to detect more cases where we can
statically determine whether or not a branch is taken.

This currently doesn't have `try` or (`async`) `with` statements.

For match statements we currently set `BasicBlock::stmts` to the entire
match statement for each basic block, even though we're only interested
in one of its cases.

Within match statements binding to named patterns is currently not
handled. Similarly to wildcard they should be considered to be always
taken (assuming no guard is present).